Add remarks field to person, and update all constructors for person
This makes Remark be part of the Person class, and update all the
references of the Person constructor to be filled with an empty
remark string (default to no remarks).

Adding a new person should default to no remarks (empty string), and editing a person should retain the old remarks (unless changed by the remark command).

This also breaks loading/saving of person due to the remark field,
for now we mark it as TODO and put in a blank remark string.